New struct svec replace struct curl_slist in the global scope
Previously we used a global struct curl_slist to manage headers.
Now, given that a curl_slist is, well, a list, every time we "append"
something it goes at the end of the list.  So, if we push two times a
Content-Type header (maybe using twice the -c flag), we end up sending
two Content-Type headers.  Introducing the new svec: the new svec_add add
the header only if it wasn't already there, otherwise replace the value.
From that, generate on-demand the curl_slist.  It also keeps track
of the string passed if they are "dirty" (i.e. allocated on the heap)
or not and free(3) 'em.