struct settings now belongs only to the child
removed the verbose settings, wasn't used.  moved prompt out of
settings, in the future it *may* be possible to change it with set.
Now the settings are managed entirely by the child: they're defined in
the global header only because code in http.c expect a global variable
(will be adressed in the future.)