add interactive -i mode, and rename old -i to -A
Add a new flag to force interactive usage (i.e. always print the
prompt) using the mnemonic flag -i.  The -i was already used to skip
the authenticity of the certificates and was renamed to -A which is a
bit harder to remember and to type (and it's a good thing for such
settings).