switching to meson
from now on I'll use meson to build this project.  This is mainly due
the fact that I'd like to use various OpenBSD-specific functions, and
use a compat layer if and only if those functions are not already
available on the host we're compiling.  Makefiles alones cannot achieve
it, so we're stuck with some kind of "configure".  Given that I don't
like too much autotools (even if they're probably more popular and
portable) I'm using meson.

I will, however, consider providing a simple Makefile, maybe building
and linking always the compat layer, or investigate other options (such
as oconfigure).

In addition to this, even the old makefile that would always build the
compat layer, wasn't really a POSIX makefile I fear.